Slett ubrukt kode og flytt validering inn i den nye validerings-filen
Er usikker på om denne valideringen faktisk skal være med nå, så må høre med Eivind
