+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/behandlingsresultat/BehandlingsresultatValideringUtils.kt
@@ -1,11 +1,15 @@
 package no.nav.familie.ba.sak.kjerne.behandlingsresultat
 
 import no.nav.familie.ba.sak.common.Feil
+import no.nav.familie.ba.sak.common.FunksjonellFeil
+import no.nav.familie.ba.sak.kjerne.behandling.domene.Behandling
+import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingType
+import no.nav.familie.ba.sak.kjerne.behandling.domene.Behandlingsresultat
+import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingÅrsak
 import no.nav.familie.ba.sak.kjerne.personident.Aktør
 import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.PersonResultat
 
 object BehandlingsresultatValideringUtils {
-
     internal fun validerAtBarePersonerFremstiltKravForEllerSøkerHarFåttEksplisittAvslag(
         personerFremstiltKravFor: List<Aktør>,
         personResultater: Set<PersonResultat>
@@ -17,6 +21,38 @@ object BehandlingsresultatValideringUtils {
                 frontendFeilmelding = "Det eksisterer personer som har fått eksplisitt avslag, men som det ikke er blitt fremstilt krav for.",
                 message = "Det eksisterer personer som har fått eksplisitt avslag, men som det ikke har blitt fremstilt krav for."
             )
+        }
+    }
+
+    // TODO: Finne ut om denne kan fjernes eller skal være med
+    internal fun validerBehandlingsresultat(behandling: Behandling, resultat: Behandlingsresultat) {
+        if ((
+            behandling.type == BehandlingType.FØRSTEGANGSBEHANDLING && setOf(
+                    Behandlingsresultat.AVSLÅTT_OG_OPPHØRT,
+                    Behandlingsresultat.ENDRET_UTBETALING,
+                    Behandlingsresultat.ENDRET_UTEN_UTBETALING,
+                    Behandlingsresultat.ENDRET_OG_OPPHØRT,
+                    Behandlingsresultat.OPPHØRT,
+                    Behandlingsresultat.FORTSATT_INNVILGET,
+                    Behandlingsresultat.IKKE_VURDERT
+                ).contains(resultat)
+            ) ||
+            (behandling.type == BehandlingType.REVURDERING && resultat == Behandlingsresultat.IKKE_VURDERT)
+        ) {
+            val feilmelding = "Behandlingsresultatet ${resultat.displayName.lowercase()} " +
+                "er ugyldig i kombinasjon med behandlingstype '${behandling.type.visningsnavn}'."
+            throw FunksjonellFeil(frontendFeilmelding = feilmelding, melding = feilmelding)
+        }
+        if (behandling.opprettetÅrsak == BehandlingÅrsak.KLAGE && setOf(
+                Behandlingsresultat.AVSLÅTT_OG_OPPHØRT,
+                Behandlingsresultat.AVSLÅTT_ENDRET_OG_OPPHØRT,
+                Behandlingsresultat.AVSLÅTT_OG_ENDRET,
+                Behandlingsresultat.AVSLÅTT
+            ).contains(resultat)
+        ) {
+            val feilmelding = "Behandlingsårsak ${behandling.opprettetÅrsak.visningsnavn.lowercase()} " +
+                "er ugyldig i kombinasjon med resultat '${resultat.displayName.lowercase()}'."
+            throw FunksjonellFeil(frontendFeilmelding = feilmelding, melding = feilmelding)
         }
     }
 }
